% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brute force search.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force search}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{data frame with two numeric variables v and w (i.e the value and weight for each item).}

\item{W}{numeric variable indicating the knapsack size.}

\item{parallel}{logical condition for parallel computing.}
}
\value{
A list of the maximum knapsack value and the positions of the elements, i.e. items.
}
\description{
The brute-force search goes through all possible alternatives and return the maximum value found. This approach is of complexity O(2^n) since all possible combinations 2^n needs to be evaluated.
}
\examples{
library(parallel)
set.seed(42)
n <- 2000
knapsack_objects <- data.frame(w = sample(1:4000, size = n, replace = TRUE), v = runif(n = n, 0, 10000))

brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 3500)
brute_force_knapsack(x = knapsack_objects[1:8,], W = 2000)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 2000)


}
