% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Heuristic algorithm}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{data frame with two numeric variables v and w (i.e the value and weight for each item).}

\item{W}{numeric variable indicating the knapsack size.}
}
\value{
A list of the maximum knapsack value and the positions of the elements, i.e. items.
}
\description{
The idea behind a greedy algorithm is that in each step you choose the best item. The best item is here defined as the item with the highest value/weight ratio, chosen from the items not already packed. (One item can only be packed once.) You will add items one at a time until the allowed total weight is reached. If the total allowed weight is 1000 and the first item you pack has a weight of 400, the next item needs to have a weight of 600 or less to be allowed in the knapsack, etc.
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <- data.frame(w = sample(1:4000, size = n, replace = TRUE), v = runif(n = n, 0, 10000))

greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
greedy_knapsack(x = knapsack_objects[1:1200,], W = 2000)

}
