% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silly_name.R
\name{named_df}
\alias{named_df}
\title{Name an empty data_frame}
\usage{
named_df(name)
}
\arguments{
\item{name}{}
}
\value{
empty data_frame
}
\description{
Name an empty data_frame
}
\examples{
named_df("foo")
}
