\name{test_function}
\alias{test_function}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A function that doubles the input
}
\description{
Doubles the input x and returns it.
}
\usage{
test_function(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
a number
}
}
\value{
double the input x
}
\examples{
test_function(2)
}
