\name{testPackage-package}
\alias{testPackage-package}
\alias{testPackage}
\docType{package}
\title{
\packageTitle{testPackage}
}
\description{
\packageDescription{testPackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{testPackage}
\packageIndices{testPackage}
A simple package for testing package building.
}
\author{
\packageAuthor{testPackage}

Maintainer: \packageMaintainer{testPackage}
}
\keyword{ package }
\examples{
test_function(2)
}
